Fix JSON encode bug
Previous code in repository.py got the current git branch name by
doing a `check_output` call without the universal_newlines flag, so
it returned a byte string. Later when the request is made to
coveralls API with the branch name, the JSON encode fails with a
TypeError on the byte string.

I'm not sure why this isn't broken on TravisCI, but I'd guess it's
related to some language/encoding env vars.